Refactor SeasonCalculator, clarify Christmas/Epiphany weekday readings, and document fixed day readings
We discovered significant duplication in the SeasonCalculator and have attempted to simplify and clarify it. This commit includes a REFACTORING_DECISIONS doc that outlines the decisions we made merging two functions into one simpler function. Unfortunately, that refactor has caused a number of tests to fail, so we’ve fixed all but one of those failures. While fixing those failures, we discovered that some of the test expectations were wrong (around expecting weekday_readings on certain dates) and have updated the tests. We also discovered why 'Christmas' weekday readings were 'missing', documented that, and added a plan to support fixed day readings that run during this period in our REFACTORING PLAN. We have yet to redo the unit tests for the new SeasonCalculator so we expect they’d all be failing as we’ve totally changed the workings of it.

Test Results (tests/test_feasts.py):

+-------------------------------+
| Test Name                    | Result |
+-------------------------------+
| test_christmas_feasts         |  ok    |
| test_easter_feasts            |  ok    |
| test_feast_types              |  ok    |
| test_feast_urls               |  ok    |
| test_all_week_calculation...  |  ok    |
| test_feast_day_readings       |  ok    |
| test_season_reading_coverage  |  ok    |
| test_sunday_cycle_coverage    |  ok    |
| test_weekday_cycle_coverage   | FAIL   |
| test_weekday_vs_sunday...     |  ok    |
| test_sunday_cycle             |  ok    |
| test_weekday_cycle            |  ok    |
| test_additional_edge_cases    |  ok    |
| ... (see log for all 34)      |        |
+-------------------------------+

Summary: 33 passed, 1 failed (Pre-Lent/Ordinary Time weekday cycle 2).